Refactor sample to use top level statements
