+++ samples/ComputeSharp.Sample.Shared/ComputeSharp.Sample.Shared.projitems
@@ -9,6 +9,7 @@
     <Import_RootNamespace>ComputeSharp.Sample.Shared</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)Helpers\Formatting.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Program.cs" />
   </ItemGroup>
 </Project>
